moving some logic from card file into new hand_analysis mod
